Add a container to allow wp-cli over ssh (#26)
* Add a container to allow wp-cli over ssh

This will allow the migration script (in a different project) to use wp-cli remotely.

e.g. wp --ssh=ssh-user@host.docker.internal:2222/var/www/html db check

It cannot use dory because it's not http traffic. So it exposes the port to the host, and ssh commands for other containers use host.docker.internal

* Improve ssh container security.

Rename to local-ssh as a reminder to only use locally.
Use env variable LOCAL_SSH_PASSWORD for the password.

* Update Makefile with migrate command