CDPT-1705 Purge cache after content updates (amends)
This PR removes the `$real_scheme` part of the nginx cache key.

This is because http and https responses are the same, and we do not serve http traffic.

It will allow us to clear cache bymaking a http request to the purge-cach endpoint, from an internal network.

The same thing has been done with the intranet's cache keys.

fastcgi_cache_key "$request_method$host$request_uri:$cookie_dw_agency";